+++ tests/test_data/test_data.py
@@ -15,7 +15,6 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-import os
 from pathlib import Path
 
 from dask.array import Array