+++ CHANGELOG.rst
@@ -18,11 +18,13 @@ Unreleased
 
 Added
 -----
+- Dependency on RosettaSciIO for read/write of some file formats.
+  (`#? <https://github.com/pyxem/kikuchipy/pull/?>`_)
 
 Changed
 -------
 - Minimum Python version is now 3.10.
-  (`#? <https://github.com/pyxem/kikuchipy/pull/?>`_)
+  (`#689 <https://github.com/pyxem/kikuchipy/pull/689>`_)
 
 Removed
 -------